requirements.txt
flask
flask-mail
torch
transformers
langchain
langchain_openai
python-dotenv